+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -9,18 +10,27 @@ module Test.Cardano.Ledger.Conway.Imp (
 ) where
 
 import Cardano.Ledger.Conway.Governance (ConwayGovState, GovState)
+import Cardano.Ledger.Conway.Rules (
+  ConwayGOV,
+  ConwayLEDGER,
+ )
+import Cardano.Ledger.Core (EraRule)
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Imp.EnactSpec as Enact
 import qualified Test.Cardano.Ledger.Conway.Imp.EpochSpec as Epoch
-import Test.Cardano.Ledger.Conway.ImpTest (ConwayEraImp)
+import qualified Test.Cardano.Ledger.Conway.Imp.GovSpec as Gov
+import Test.Cardano.Ledger.Conway.ImpTest (ConwayEraImp, withImpState)
 
 spec ::
   forall era.
   ( ConwayEraImp era
   , GovState era ~ ConwayGovState era
+  , EraRule "GOV" era ~ ConwayGOV era
+  , EraRule "LEDGER" era ~ ConwayLEDGER era
   ) =>
   Spec
 spec = do
-  describe "ConwayImpSpec" $ do
+  describe "ConwayImpSpec" $ withImpState @era $ do
     Enact.spec @era
     Epoch.spec @era
+    Gov.spec @era